backend/requirements_secure.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Session==0.8.0
python-dotenv==1.0.0
google-cloud-firestore==2.16.0
gunicorn==21.2.0
requests==2.32.3
PyJWT==2.9.0
cryptography==42.0.8